Create Reversed Words
i ripped my head open with this pattern of split, reverse, join methods that i sensed the solution right away. god i feel so lame that i can barely understand this after such a long time. but now that i think of it, even during progress my brain finds a way to hurt me. fuck that. i will pat myself on the back for this