Create  The Wide-Mouthed frog!
this was the first time i used the PREP method to solve a coding challenge. this made me pause and take inventory of my thinking. i also had to ponder the steps needed to solve it. 

i thought it was a waste of time doing PREP for easier coding challenges but i was so wrong. doing it for the easy ones sets me up for being able to PREP my way through the hard ones !