Create anagram check
my first time going at a 7 kyu with the PREP method. i can see that i was still thinking that an 8kyu approack would solve this by adding the includes method to 'test' and calling it a day. but i was dead wrong. i was stuck on how to compare individual elements in each parameter, but looking at someone else's solution totally blindsided me in that i actually knew how to read what was going on. 

i was able to understand what was going on becuase i did the PREP method, or as far as i could go with it