Create Alternate capitalization
i was trying to use the filter method to solve this, but after hitting my head against the desk for 15 minutes, i looked at the solution and saw that the map method was the right method to use. i still dont get this solution but that's tomorrow's problem. time to go about my day