Create Who likes it?
i toyed with map and for each methods but the output needed to be a string and not an array. i then knew i had to write the logic for each case, for 0-3 strings, and then 4+ strings

i opted to write if/else statements along with template literals. other solutions point towards switch/case statements. i definately need to practice switch/case statements since i am not as familiar with them as if/else, but i got the job done here so yay