Create Mexican Wave
i got lost in how to solve this. i understood what to do, even trying to use a loop. but now i see that i should have used the split method first, then looped through the array--setting up a conditional to look for spaces--and pushing capitalized letters into a new array, as well as using the join method