Create Credit Card Mask
i was totally lost in this one. i had vague ideas of using slice and replace methods, and was unsure of how to go about it using loops. this solution made the most sense to me since it used a loop, other solutions using regex (not familiar enough with that yet)