Create Find the odd int
i have a wiff of how to do this problem but no idea exactly how to do it. this loop solution was what my intuition was telling me to do, and i now somewhat grasp what this solution is doing. AI alluded to objects being more optimal but let's save that head splitter for another day